fix errors and warnings regarding RNG
The following warning was fixed: utils.cpp:169:23: warning: narrowing conversion of ‘seed’ from ‘int’ to ‘std::mersenne_twister_engine<long unsigned int, 32, 624, 397, 31, 2567483615, 11, 4294967295, 7, 2636928640, 15, 4022730752, 18, 1812433253>::result_type’ {aka ‘long unsigned int’} [-Wnarrowing]
  169 |     twister = mt19937{seed};

The following bug was fixed: previously, there was a INT8_MAX cap to seed generation, which was too short. The new cap of ULONG_MAX is better.

+++ main.cpp
@@ -3,6 +3,7 @@
 #include "graph.hpp"
 #include "utils.hpp"
 #include "params.hpp"
+#include <climits>
 #include <fstream>
 #include <iostream>
 #include <vector>
@@ -21,7 +22,7 @@ int main(int argc, char *argv[]) {
     const int samples = param_samples();
     // Initial seed to set a different and random seed for each sample
     set_seed(param_seed());
-    const vector<int> samples_seeds = rand_vector(samples, INT8_MAX);
+    const vector<long unsigned int> samples_seeds = rand_vector(samples, ULONG_MAX);
     const int N = param_N();
     const int T = param_T();
     const int BURN_T = param_BURN_T();
